fixed max score behaviour

+++ Assets/Scripts/Controller/Actions/AppendMaxScore.cs
@@ -8,8 +8,8 @@ namespace Controller
 	{
 		override public PrefromResult Perform (float delta)
 		{
-			if (GameModel.Instance().maxScore > PlayerPrefs.GetInt ("highscore", 0))
-				PlayerPrefs.SetInt ("highscore", GameModel.Instance().maxScore);
+			if ((int)GameModel.Instance().maxScore > PlayerPrefs.GetInt ("highscore", 0))
+				PlayerPrefs.SetInt ("highscore", (int)GameModel.Instance().maxScore);
 
 			return PrefromResult.COMPLETED;
 		}
